Fix udp

+++ ssh/udp-custom.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 cd
+if [ -d /etc/udp ];then
 rm -rf /etc/udp
+fi
 mkdir -p /etc/udp
 
 # change to time GMT+7
@@ -14,41 +16,26 @@ wget -q -O /etc/udp/udp-custom "https://raw.githubusercontent.com/zhets/project/
 chmod +x /etc/udp/udp-custom
 echo downloading default config
 wget -q -O /etc/udp/config.json "https://raw.githubusercontent.com/zhets/project/main/ssh/config.json"
-chmod 644 /etc/udp/config.json
+chmod 777 /etc/udp/config.json
 
-if [ -z "$1" ]; then
-cat <<EOF > /etc/systemd/system/udp-custom.service
+cat> /etc/systemd/system/udp-custom.service <<-END
 [Unit]
-Description=UDP Custom by ePro Dev. Team
+Description=SSH UDP Custom Service
+Documentation=https://t.me/xd_tunnel
+After=network.target nss-lookup.target
 
 [Service]
 User=root
 Type=simple
-ExecStart=/etc/udp/udp-custom server
+ExecStart=/etc/udp/udp-custom server -exclude 1,54,55,1000,65535
 WorkingDirectory=/etc/udp/
 Restart=always
-RestartSec=2s
+RestartSec=5s
 
 [Install]
 WantedBy=default.target
-EOF
-else
-cat <<EOF > /etc/systemd/system/udp-custom.service
-[Unit]
-Description=UDP Custom by ePro Dev. Team
+END
 
-[Service]
-User=root
-Type=simple
-ExecStart=/etc/udp/udp-custom server -exclude $1
-WorkingDirectory=/etc/udp/
-Restart=always
-RestartSec=2s
-
-[Install]
-WantedBy=default.target
-EOF
-fi
 
 echo start service udp-custom
 systemctl start udp-custom &>/dev/null
